+++ src/Infrastructure/Persistence/UnitOfWorkManager.php
@@ -20,7 +20,10 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Infrastructure\Persistence;
 
 use Closure;
+use CloudCreativity\Modules\Infrastructure\Infrastructure;
 use CloudCreativity\Modules\Infrastructure\InfrastructureException;
+use CloudCreativity\Modules\Infrastructure\Log\ExceptionReporterInterface;
+use Throwable;
 
 final class UnitOfWorkManager implements UnitOfWorkManagerInterface
 {
@@ -48,9 +51,12 @@ final class UnitOfWorkManager implements UnitOfWorkManagerInterface
      * UnitOfWorkManager constructor.
      *
      * @param UnitOfWorkInterface $unitOfWork
+     * @param ExceptionReporterInterface|null $reporter
      */
-    public function __construct(private readonly UnitOfWorkInterface $unitOfWork)
-    {
+    public function __construct(
+        private readonly UnitOfWorkInterface $unitOfWork,
+        private readonly ?ExceptionReporterInterface $reporter = null,
+    ) {
     }
 
     /**
@@ -58,19 +64,49 @@ final class UnitOfWorkManager implements UnitOfWorkManagerInterface
      */
     public function execute(Closure $callback, int $attempts = 1): mixed
     {
-        if ($this->active) {
-            throw new InfrastructureException(
-                'Not expecting unit of work manager to start a unit of work within an existing one.',
-            );
+        Infrastructure::assert($attempts > 0, 'Attempts must be greater than zero.');
+        Infrastructure::assert(
+            $this->active === false,
+            'Not expecting unit of work manager to start a unit of work within an existing one.',
+        );
+
+        return $this->retry($callback, $attempts);
+    }
+
+    /**
+     * @param Closure $callback
+     * @param int $attempts
+     * @return mixed
+     */
+    private function retry(Closure $callback, int $attempts): mixed
+    {
+        try {
+            return $this->transaction($callback);
+        } catch (Throwable $ex) {
+            if ($attempts === 1) {
+                throw $ex;
+            }
+
+            // Report "swallowed" exceptions.
+            $this->reporter?->report($ex);
         }
 
+        return $this->retry($callback, $attempts - 1);
+    }
+
+    /**
+     * @param Closure $callback
+     * @return mixed
+     */
+    private function transaction(Closure $callback): mixed
+    {
         try {
             $result = $this->unitOfWork->execute(function () use ($callback) {
                 $this->active = true;
                 $value = $callback();
                 $this->executeBeforeCommit();
                 return $value;
-            }, $attempts);
+            });
             $this->committed = true;
             $this->executeAfterCommit();
             return $result;