+++ tests/Unit/Infrastructure/Persistence/UnitOfWorkManagerTest.php
@@ -19,9 +19,12 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Tests\Unit\Infrastructure\Persistence;
 
+use Closure;
 use CloudCreativity\Modules\Infrastructure\InfrastructureException;
+use CloudCreativity\Modules\Infrastructure\Log\ExceptionReporterInterface;
 use CloudCreativity\Modules\Infrastructure\Persistence\UnitOfWorkInterface;
 use CloudCreativity\Modules\Infrastructure\Persistence\UnitOfWorkManager;
+use LogicException;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
@@ -31,6 +34,11 @@ class UnitOfWorkManagerTest extends TestCase
      * @var UnitOfWorkInterface&MockObject
      */
     private UnitOfWorkInterface&MockObject $unitOfWork;
+
+    /**
+     * @var MockObject&ExceptionReporterInterface
+     */
+    private ExceptionReporterInterface&MockObject $reporter;
 
     /**
      * @var UnitOfWorkManager
@@ -46,6 +54,7 @@ class UnitOfWorkManagerTest extends TestCase
 
         $this->manager = new UnitOfWorkManager(
             $this->unitOfWork = $this->createMock(UnitOfWorkInterface::class),
+            $this->reporter = $this->createMock(ExceptionReporterInterface::class),
         );
     }
 
@@ -59,8 +68,8 @@ class UnitOfWorkManagerTest extends TestCase
         $this->unitOfWork
             ->expects($this->once())
             ->method('execute')
-            ->willReturnCallback(function (\Closure $callback, int $attempts) use (&$sequence) {
-                $this->assertSame(2, $attempts);
+            ->willReturnCallback(function (Closure $callback, int $attempts) use (&$sequence) {
+                $this->assertSame(1, $attempts);
                 $sequence[] = 'start';
                 $result = $callback();
                 $sequence[] = 'commit';
@@ -112,7 +121,7 @@ class UnitOfWorkManagerTest extends TestCase
         $this->unitOfWork
             ->expects($this->once())
             ->method('execute')
-            ->willReturnCallback(function (\Closure $callback, int $attempts) use (&$sequence, $expected) {
+            ->willReturnCallback(function (Closure $callback, int $attempts) use (&$sequence, $expected) {
                 $this->assertSame(1, $attempts);
                 $sequence[] = 'start';
                 $callback();
@@ -159,8 +168,8 @@ class UnitOfWorkManagerTest extends TestCase
         $this->unitOfWork
             ->expects($this->once())
             ->method('execute')
-            ->willReturnCallback(function (\Closure $callback, int $attempts) use (&$sequence) {
-                $this->assertSame(2, $attempts);
+            ->willReturnCallback(function (Closure $callback, int $attempts) use (&$sequence) {
+                $this->assertSame(1, $attempts);
                 $sequence[] = 'start';
                 $result = $callback();
                 $sequence[] = 'commit';
@@ -236,7 +245,7 @@ class UnitOfWorkManagerTest extends TestCase
     {
         $this->unitOfWork
             ->method('execute')
-            ->willReturnCallback(fn (\Closure $callback) => $callback());
+            ->willReturnCallback(fn (Closure $callback) => $callback());
 
         $result1 = $this->manager->execute(function () {
             $this->manager->beforeCommit(fn () => null);
@@ -256,14 +265,22 @@ class UnitOfWorkManagerTest extends TestCase
      */
     public function testItFlushesCallbacksOnUnsuccessfulTransaction(): void
     {
-        $expected = new \RuntimeException('Boom');
+        $ex = new LogicException('Boom');
+        $count = 0;
+
+        $this->reporter
+            ->expects($this->exactly(2)) // only "swallowed" exceptions are reported
+            ->method('report')
+            ->with($this->identicalTo($ex));
 
         $this->unitOfWork
+            ->expects($this->exactly(4))
             ->method('execute')
-            ->willReturnCallback(function (\Closure $callback, int $attempts) use ($expected) {
+            ->willReturnCallback(function (\Closure $callback) use ($ex, &$count) {
+                $count++;
                 $result = $callback();
-                if (1 === $attempts) {
-                    throw $expected;
+                if ($count < 4) {
+                    throw $ex;
                 }
                 return $result;
             });
@@ -272,16 +289,191 @@ class UnitOfWorkManagerTest extends TestCase
             $this->manager->execute(function () {
                 $this->manager->beforeCommit(fn () => null);
                 $this->manager->afterCommit(fn () => null);
-            });
+            }, 2);
             $this->fail('No exception thrown.');
-        } catch (\RuntimeException $ex) {
-            $this->assertSame($expected, $ex);
+        } catch (LogicException $ex) {
+            $this->assertSame($ex, $ex);
         }
 
         // this would error if the manager is not empty at this point (tests below prove that).
         $result = $this->manager->execute(fn () => true, 2);
 
         $this->assertTrue($result);
+    }
+
+    /**
+     * Scenario in which the callback (not the unit of work) errors, with re-attempts.
+     *
+     * In this scenario, no before or after commit callbacks should be executed. This is
+     * because the callback is executed (and throws) before the inner unit of work is
+     * committed. Therefore, no before or after callbacks can be executed.
+     *
+     * However, in this scenario any callbacks that were registered before the exception
+     * is thrown should be forgotten. Otherwise, if the callback is retried, the before
+     * callbacks will be executed twice.
+     *
+     * @return void
+     */
+    public function testItHandlesCallbackExecutingMultipleTimes(): void
+    {
+        $ex = new LogicException('Boom!');
+        $sequence = [];
+        $count = 0;
+
+        $this->reporter
+            ->expects($this->exactly(2))
+            ->method('report')
+            ->with($this->identicalTo($ex));
+
+        $before = [
+            function () use (&$sequence): void {
+                $sequence[] = 'before1';
+            },
+            function () use (&$sequence): void {
+                $sequence[] = 'before2';
+            },
+            function () use (&$sequence): void {
+                $sequence[] = 'before3';
+            },
+        ];
+
+        $after = [
+            function () use (&$sequence): void {
+                $sequence[] = 'after1';
+            },
+            function () use (&$sequence): void {
+                $sequence[] = 'after2';
+            },
+        ];
+
+        $this->unitOfWork
+            ->expects($this->exactly(3))
+            ->method('execute')
+            ->willReturnCallback(function (Closure $closure) use (&$sequence): mixed {
+                try {
+                    $sequence[] = 'begin';
+                    $result = $closure();
+                    $sequence[] = 'committed';
+                    return $result;
+                } catch (\Throwable $ex) {
+                    $sequence[] = 'rollback';
+                    throw $ex;
+                }
+            });
+
+        $this->manager->execute(function () use ($before, $after, $ex, &$count): void {
+            $count++;
+            $this->manager->afterCommit($after[0]);
+            $this->manager->beforeCommit($before[0]);
+            $this->manager->beforeCommit($before[1]);
+
+            if ($count < 3) {
+                throw $ex;
+            }
+
+            $this->manager->beforeCommit($before[2]);
+            $this->manager->afterCommit($after[1]);
+        }, 3);
+
+        $this->assertSame([
+            'begin',
+            'rollback',
+            'begin',
+            'rollback',
+            'begin',
+            'before1',
+            'before2',
+            'before3',
+            'committed',
+            'after1',
+            'after2',
+        ], $sequence);
+    }
+
+    /**
+     * Scenario where the inner unit of work fails several times.
+     *
+     * In this scenario, any before callbacks will be executed multiple times. This
+     * is because the manager executes them before the inner unit of work commits
+     * (and throws). Therefore, the before callbacks are executed again when the
+     * callback is retried and is successful.
+     *
+     * However, after commit callbacks should not be executed until the successful
+     * commit of the inner transaction. They should not be executed more than once,
+     * because any that are registered on one of the failed attempts should havbe been
+     * forgotten.
+     *
+     * @return void
+     */
+    public function testItHandlesUnitOfWorkFailingMultipleTimes(): void
+    {
+        $ex = new LogicException('Boom!');
+        $sequence = [];
+        $count = 0;
+
+        $this->reporter
+            ->expects($this->exactly(2))
+            ->method('report')
+            ->with($this->identicalTo($ex));
+
+        $before = [
+            function () use (&$sequence): void {
+                $sequence[] = 'before1';
+            },
+            function () use (&$sequence): void {
+                $sequence[] = 'before2';
+            },
+        ];
+
+        $after = [
+            function () use (&$sequence): void {
+                $sequence[] = 'after1';
+            },
+            function () use (&$sequence): void {
+                $sequence[] = 'after2';
+            },
+        ];
+
+        $this->unitOfWork
+            ->method('execute')
+            ->willReturnCallback(function (Closure $closure) use ($ex, &$sequence, &$count): mixed {
+                $count++;
+
+                $sequence[] = 'begin';
+                $result = $closure();
+
+                if ($count < 3) {
+                    $sequence[] = 'rollback';
+                    throw $ex;
+                }
+
+                $sequence[] = 'committed';
+                return $result;
+            });
+
+        $this->manager->execute(function () use ($before, $after): void {
+            $this->manager->afterCommit($after[0]);
+            $this->manager->beforeCommit($before[0]);
+            $this->manager->beforeCommit($before[1]);
+            $this->manager->afterCommit($after[1]);
+        }, 3);
+
+        $this->assertSame([
+            'begin',
+            'before1',
+            'before2',
+            'rollback',
+            'begin',
+            'before1',
+            'before2',
+            'rollback',
+            'begin',
+            'before1',
+            'before2',
+            'committed',
+            'after1',
+            'after2',
+        ], $sequence);
     }
 
     /**
@@ -318,7 +510,7 @@ class UnitOfWorkManagerTest extends TestCase
         $this->unitOfWork
             ->expects($this->once())
             ->method('execute')
-            ->willReturnCallback(fn (\Closure $callback) => $callback());
+            ->willReturnCallback(fn (Closure $callback) => $callback());
 
         $this->expectException(InfrastructureException::class);
         $this->expectExceptionMessage(
@@ -337,7 +529,7 @@ class UnitOfWorkManagerTest extends TestCase
     {
         $this->unitOfWork
             ->method('execute')
-            ->willReturnCallback(fn (\Closure $callback) => $callback());
+            ->willReturnCallback(fn (Closure $callback) => $callback());
 
         $this->expectException(InfrastructureException::class);
         $this->expectExceptionMessage('Cannot queue a before commit callback as unit of work has been committed.');